refactor: make the requested changes per review
