+++ .vscode/settings.json
@@ -9,5 +9,5 @@
   "[typescriptreact]": {
     "editor.defaultFormatter": "esbenp.prettier-vscode"
   },
-  "cSpell.words": ["Hajdin", "hoverable"]
+  "cSpell.words": ["Hajdin", "hoverable", "nextui"]
 }
