+++ components/workExperience/Companies.tsx
@@ -1,0 +1,20 @@
+"use client"; // NOTE - companies is a client-side svg component
+
+import { companies } from "../icons";
+import Company from "./Company";
+
+const Companies = () => {
+  return (
+    <>
+      {companies.map((CompanyLogo, index) => (
+        <Company
+          key={CompanyLogo.name}
+          CompanyLogo={CompanyLogo}
+          index={index}
+        />
+      ))}
+    </>
+  );
+};
+
+export default Companies;
