+++ styles/globals.css
@@ -102,7 +102,7 @@
   box-shadow: 10px 10px 26px 0px rgba(2, 82, 205, 0.28);
 }
 
-.mySkills {
+.mobileHeading2 {
   color: var(--black-200, #151e2c);
   font-size: 2.25rem;
   font-style: normal;
